test(openai): move model override coverage to core owner
