+++ src/agents/embedded-agent-runner/model.test.ts
@@ -4,6 +4,7 @@ import os from "node:os";
 import path from "node:path";
 import { MAX_TIMER_TIMEOUT_MS } from "@openclaw/normalization-core/number-coercion";
 import { beforeEach, describe, expect, it, vi } from "vitest";
+import { buildOpenAIProvider } from "../../../extensions/openai/api.js";
 import { withEnvAsync } from "../../test-utils/env.js";
 import { discoverAuthStorage, discoverModels } from "../agent-model-discovery.js";
 import {
@@ -259,6 +260,7 @@ import type { ModelDefinitionConfig, ModelProviderConfig } from "../../config/ty
 import type { Model } from "../../llm/types.js";
 import { getModelProviderLocalService } from "../provider-local-service.js";
 import { getModelProviderRequestTransport } from "../provider-request-config.js";
+import { applyConfiguredProviderOverrides } from "./model.configured-overrides.js";
 import { buildForwardCompatTemplate } from "./model.forward-compat.test-support.js";
 import { buildInlineProviderModels } from "./model.inline-provider.js";
 import { resolveModelAsync, resolveModelWithRegistry } from "./model.js";
@@ -4477,6 +4479,77 @@ describe("resolveModel", () => {
     expect(result.error).toBe(
       "Unknown model: openai/gpt-5.3-codex-spark. gpt-5.3-codex-spark is available only through ChatGPT/Codex OAuth. Run `openclaw models auth login --provider openai` and use openai/gpt-5.3-codex-spark with that OAuth profile; OpenAI API-key auth cannot use this model.",
     );
+  });
+
+  it.each(["provider", "model"])("preserves authored %s transport and model overrides", (scope) => {
+    const provider = buildOpenAIProvider();
+    const modelId = "gpt-5.6-luna";
+    const route = { api: "openai-completions", baseUrl: "https://proxy.example/v1" } as const;
+    const providerConfig: ModelProviderConfig = {
+      baseUrl: "https://api.openai.com/v1",
+      ...(scope === "provider" ? route : {}),
+      headers: { "X-Provider-Route": "authored" },
+      models: [
+        {
+          id: modelId,
+          name: "Authored Luna",
+          ...(scope === "model" ? route : {}),
+          headers: { "X-Model-Route": "authored" },
+          compat: { codeMode: "capable", supportsTemperature: true },
+          reasoning: false,
+          input: ["text"],
+          contextWindow: 64_000,
+          maxTokens: 4_000,
+          cost: { input: 2, output: 3, cacheRead: 0, cacheWrite: 0 },
+        },
+      ],
+    };
+    const config = { models: { providers: { openai: providerConfig } } };
+    const discoveredModel = provider.resolveDynamicModel?.({
+      provider: "openai",
+      modelId,
+      providerConfig,
+      config,
+      modelRegistry: {
+        find: () => undefined,
+        getAll: () => [],
+        getAvailable: () => [],
+        hasConfiguredAuth: () => false,
+      },
+    });
+    if (!discoveredModel) {
+      throw new Error("expected the OpenAI dynamic model");
+    }
+    expect(discoveredModel.compat?.codeMode).toBe("preferred");
+
+    const model = applyConfiguredProviderOverrides({
+      provider: "openai",
+      modelId,
+      discoveredModel,
+      providerConfig,
+      cfg: config,
+      manifestAlias: { provider: "openai" },
+      runtimeHooks: {
+        buildProviderUnknownModelHintWithPlugin: ({ context }) =>
+          provider.buildUnknownModelHint?.(context) ?? undefined,
+        prepareProviderDynamicModel: async () => undefined,
+        runProviderDynamicModel: ({ context }) => provider.resolveDynamicModel?.(context),
+        normalizeProviderResolvedModelWithPlugin: ({ context }) =>
+          provider.normalizeResolvedModel?.(context),
+        normalizeProviderTransportWithPlugin: ({ context }) =>
+          provider.normalizeTransport?.(context) ?? undefined,
+      },
+    });
+    expect(model).toMatchObject({
+      ...route,
+      headers: { "X-Provider-Route": "authored", "X-Model-Route": "authored" },
+      compat: { codeMode: "capable", supportsTemperature: true },
+      reasoning: false,
+      input: ["text"],
+      contextWindow: 64_000,
+      maxTokens: 4_000,
+      cost: { input: 2, output: 3, cacheRead: 0, cacheWrite: 0 },
+    });
   });
 
   it("applies provider overrides to openai gpt-5.4 forward-compat models", async () => {
